Change theme tags' behaviour.
In the database, theme tags are assigned to english language only.
When generating JSON, however, non-theme tags must be taken into
account. If non-theme tags are present in a language, its JSON
will be generated. Otherwise, it won't be.
This creates a problem because themes will still be in english
language only. Now if english has no non-theme tags, its theme tags
will also be lost. In cases like these, we move the theme tags from
english to another language, so that english can then be safely
ignored. The assumption here is that in such cases, only one
(non-english) language will have non-theme tags, and that's the
language which will also get the theme tags while generating JSON.
Note that this will not change the database. In the database,
theme tags will still remain assigned to english language only.